Moved debug functions to their own file (#2740)
